requirements.txt
flask
flask-dotenv
requests
beautifulsoup4
openai
